Improved pulse wave response at higher octaves -softens the edges of pulse waves at high frequencies to reduce aliasing noise -Rearranged the order of the oscillators in the UI. -Oscillator.cs:  Brown noise no longer pops when initializing thanks to saner init value -PinkNoise.cs:  P_URand can now take a negative value to set the seed to "Now" minus an offset. -EGPanel:  Can click the waveform for a "fast selection" popup.  Slider still exists to allow scrubbing.

+++ FMCore/Voice.cs
@@ -172,7 +172,7 @@ namespace gdsFM
         }
 
         /// Sets the canonical waveform to reference when setting an operator's waveFunc on NoteOn.
-        public void SetWaveform(int opTarget, float val)
+        public void SetWaveform(int opTarget, int val)
         {   // NOTE:  This does NOT actually set an operator's waveFunc!  This is done in NoteOn when referencing this value from Voice.
             opType[opTarget] = (byte)val;
         }
